Add interactive half-star rating on event cards
- Migrate event.rating from INTEGER to FLOAT for 0.5 increments
- Interactive StarRating component: hover shows preview, left half = half star
- Stars always visible on cards; empty outline if unrated
- Clicking same rating clears it; optimistic update with API revert on failure
- patchEventRating() PATCH endpoint call in api.ts

+++ backend/app/models/core.py
@@ -62,7 +62,7 @@ class Event(SQLModel, table=True):
     work_id: Optional[uuid.UUID] = Field(default=None, foreign_key="work.id")
     price_paid: Optional[Decimal] = None
     currency: Optional[str] = None  # ISO 4217, e.g. GBP, EUR
-    rating: Optional[int] = None
+    rating: Optional[float] = None
     notes: Optional[str] = None
     festival_id: Optional[uuid.UUID] = Field(default=None, foreign_key="festival.id")
     substack_url: Optional[str] = None
