+++ backend/app/schemas/events.py
@@ -21,7 +21,7 @@ class EventBase(BaseModel):
     work_id: Optional[uuid.UUID] = None
     price_paid: Optional[Decimal] = None
     currency: Optional[str] = None
-    rating: Optional[int] = None
+    rating: Optional[float] = None
     notes: Optional[str] = None
     festival_id: Optional[uuid.UUID] = None
     substack_url: Optional[str] = None
@@ -38,7 +38,7 @@ class EventUpdate(BaseModel):
     work_id: Optional[uuid.UUID] = None
     price_paid: Optional[Decimal] = None
     currency: Optional[str] = None
-    rating: Optional[int] = None
+    rating: Optional[float] = None
     notes: Optional[str] = None
     festival_id: Optional[uuid.UUID] = None
     substack_url: Optional[str] = None
@@ -63,7 +63,7 @@ class EventListItem(BaseModel):
     festival_name: Optional[str] = None
     price_paid: Optional[Decimal] = None
     currency: Optional[str] = None
-    rating: Optional[int] = None
+    rating: Optional[float] = None
     data_completeness: Optional[str] = None
     substack_url: Optional[str] = None
     status: str = "attended"
@@ -87,7 +87,7 @@ class EventDetail(BaseModel):
     festival: Optional[NamedRef] = None
     price_paid: Optional[Decimal] = None
     currency: Optional[str] = None
-    rating: Optional[int] = None
+    rating: Optional[float] = None
     notes: Optional[str] = None
     substack_url: Optional[str] = None
     data_completeness: Optional[str] = None
